a few things added
The main attention was how to integrate the algorithm to the baseline application. I thought about tying the algo to the API and the app would send and API request with the user ID and feedback.json. But that would eb intensive. So, the next step is to tie the algo to the front end application, open the feedback.json in the backend on of the client facing app (is that possible?), make it globally available, and have the algorithm use it. once that's done and the user finishes using the app, the algo would then update the feedback.json and then push it to a db. THIS is the part where I'm confused on how to execute, but it's fine. I can figure it out.