i'm tired
i tried a BUNCH of different stuff, which is why you can't see like anything in the push. anyways, I forgot PYDANTIC was a thing and now I'm gonna focus on that...on Sunday. I'm taking Saturday off (well, i'm pushing on saturday so not really, but you get what I mean)